move meter and predictor drag and drop
