+++ src/app/upload-data/data-setup/data-setup-banner/data-setup-banner.component.html
@@ -5,7 +5,7 @@
                 [routerLinkActive]="['active']">
                 <i class="fa fa-file"></i>{{reference.name}}</a>
         </li>
-        <li class="nav-item">
+        <!-- <li class="nav-item">
             <a class="nav-link" routerLink="/upload/data-setup/set-facility-meters"
             [routerLinkActive]="['active']">
                 <i class="fa fa-industry"></i>Set Facility Meters</a>
@@ -14,7 +14,7 @@
             <a class="nav-link" routerLink="/upload/data-setup/set-facility-predictors"
             [routerLinkActive]="['active']">
                 <i class="fa fa-industry"></i>Set Facility Predictors</a>
-        </li>
+        </li> -->
         <li class="nav-item">
             <a class="nav-link" routerLink="/upload/data-setup/manage-meters" [routerLinkActive]="['active']">
                 <i class="fa fa-gauge-high"></i>Meters</a>
