+++ src/app/upload-data/data-setup/file-setup/file-setup.component.html
@@ -12,6 +12,16 @@
                     <i class="fa fa-file-invoice"></i>
                     Identify Columns</a>
             </li>
+            <li class="nav-item">
+                <a class="nav-link" routerLink="set-facility-meters"
+                [routerLinkActive]="['active']">
+                    <i class="fa fa-industry"></i>Set Facility Meters</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" routerLink="set-facility-predictors"
+                [routerLinkActive]="['active']">
+                    <i class="fa fa-industry"></i>Set Facility Predictors</a>
+            </li>
         </ul>
     </div>
