+++ src/app/upload-data/data-setup/file-setup/set-facility-meters/set-facility-meters.component.html
@@ -1,0 +1,27 @@
+<div class="d-flex justify-content-end">
+    <div class="d-flex">
+        <button class="btn btn-success" (click)="continue()">Submit &amp; Continue</button>
+    </div>
+</div>
+<h6 class="map-columns-label">Map the meters to their corresponding facilities by dragging and dropping them
+    into the correct columns.
+</h6>
+<hr>
+<div class="row no-gutters">
+    <div cdkDropList [cdkDropListData]="fileReference.meterFacilityGroups" class="card col-md-3 col-sm-6"
+        *ngFor="let facilityGroup of fileReference.meterFacilityGroups; let index = index"
+        (cdkDropListDropped)="dropColumn($event)" id={{facilityGroup.facilityId}}
+        [cdkDropListConnectedTo]="facilityGroupIds">
+        <div class="card-header">
+            <i *ngIf="index != 0" class="fa fa-industry" [ngStyle]="{color: facilityGroup.color}"></i>
+            {{facilityGroup.facilityName}}
+        </div>
+        <div class="card-body">
+            <div class="list-item mt-2 mb-2 p-2" [cdkDragData]="groupItem"
+                *ngFor="let groupItem of facilityGroup.groupItems" cdkDrag>
+                <span class="fa fa-arrows-alt"></span>{{groupItem.value}} <br>
+                {{groupItem.fileName}}
+            </div>
+        </div>
+    </div>
+</div>
