--- src/app/upload-data/data-setup/set-facility-meters/set-facility-meters.component.html
@@ -1,32 +1,0 @@
-<div class="wrapper main-content p-0">
-    <div class="content-padding">
-        <h4>Set Facility Meters</h4>
-        <div class="d-flex justify-content-end">
-            <div class="d-flex">
-                <button class="btn btn-success" (click)="continue()">Submit &amp; Continue</button>
-            </div>
-        </div>
-        <h6 class="map-columns-label">Map the meters to their corresponding facilities by dragging and dropping them
-            into the correct columns.
-        </h6>
-        <hr>
-        <div class="row no-gutters">
-            <div cdkDropList [cdkDropListData]="facilityGroups" class="card col-md-3 col-sm-6"
-                *ngFor="let facilityGroup of facilityGroups; let index = index"
-                (cdkDropListDropped)="dropColumn($event)" id={{facilityGroup.facilityId}}
-                [cdkDropListConnectedTo]="facilityGroupIds">
-                <div class="card-header">
-                    <i *ngIf="index != 0" class="fa fa-industry" [ngStyle]="{color: facilityGroup.color}"></i>
-                    {{facilityGroup.facilityName}}
-                </div>
-                <div class="card-body">
-                    <div class="list-item mt-2 mb-2 p-2" [cdkDragData]="groupItem"
-                        *ngFor="let groupItem of facilityGroup.groupItems" cdkDrag>
-                        <span class="fa fa-arrows-alt"></span>{{groupItem.value}} <br>
-                        {{groupItem.fileName}}
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
