+++ src/app/upload-data/upload-data.module.ts
@@ -11,8 +11,8 @@ import { FormsModule } from '@angular/forms';
 import { IdentifyColumnsComponent } from './data-setup/file-setup/identify-columns/identify-columns.component';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { ManageMetersComponent } from './data-setup/manage-meters/manage-meters.component';
-import { SetFacilityPredictorsComponent } from './data-setup/set-facility-predictors/set-facility-predictors.component';
-import { SetFacilityMetersComponent } from './data-setup/set-facility-meters/set-facility-meters.component';
+import { SetFacilityPredictorsComponent } from './data-setup/file-setup/set-facility-predictors/set-facility-predictors.component';
+import { SetFacilityMetersComponent } from './data-setup/file-setup/set-facility-meters/set-facility-meters.component';
